+++ actions/tests/actions_tests/test_lint_actions/actions/bad_actions.py
@@ -1,0 +1,13 @@
+from robocorp.actions import action
+
+
+@action
+def bad_action_no_return():
+    """
+    Empty...
+    """
+
+
+@action
+def bad_action_no_docstring() -> str:
+    return ""